pyproject.toml
[tool.poetry]
name = "gtfs-go"
version = "0.1.0"
description = ""
authors = ["Kanahiro <mediterranean1769@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
ruff = "^0.4.3"
pytest-qt = "^4.4.0"
pytest-cov = "^5.0.0"
pandas = "^2.2.2"
pytest-mock = "^3.14.0"
pytest-qgis = "^2.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
select = ["F", "E", "W", "B", "C90", "N", "I"]
ignore = ["N999", "N802", "F401", "E501"]
target-version = "py39"